Seclude setupServerless test util
